refactor: split non-component exports to fix react-refresh warnings
Fast Refresh only works when a file exports components exclusively. Move
the constants, hooks, and utilities that were co-located with component
exports into their own modules so each component file exports only
components, following the existing *-provider / *-hooks convention.

Also remove a dead eslint-disable directive in DrawingPad.

Resolves all 16 react-refresh/only-export-components (and unused-directive)
lint warnings.
