+++ src/components/dependent/DrawingPad.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import {
   Dispatch,
   ReactNode,
