+++ src/components/dependent/routing/router-adapter.tsx
@@ -1,4 +1,4 @@
-import { Route, Switch, useLocation, useSearch, useSearchParams } from "wouter";
+import { Route, Switch } from "wouter";
 import { RememberedHomeLink as Link } from "./remembered-home-link";
 
-export { Route, Switch, Link, useLocation, useSearch, useSearchParams };
+export { Route, Switch, Link };
