+++ src/components/dependent/routing/router-hooks.ts
@@ -1,0 +1,1 @@
+export { useLocation, useSearch, useSearchParams } from "wouter";
