+++ src/components/error/Sumimasen.tsx
@@ -1,27 +1,3 @@
-export const APOLOGIZING_SYMBOL = [
-  "(シ_ _)シ",
-  "m(_ _)m",
-  "m(._.)m",
-  "<(_ _)>",
-  "(_ _*)",
-  "m(｡≧ _ ≦｡)m",
-  //  "(´；ω；｀)",
-  "(｡•́︿•̀｡)",
-  //  "(；人；)",
-  "(;-;)",
-  "(>_<)",
-  // "(·.·)",
-  // "(^-^*)",
-  // "\\(^Д^)/",
-  // "\\(o_o)/",
-  // "(˚Δ˚)b",
-  // "(≥o≤)",
-  // "(='X'=)",
-  // "(o^^)o",
-  "(╬ Ò ‸ Ó)",
-  "( •́ ‿ •̀ )ゞ",
-];
-
 export const Sumimasen = ({
   layout = "stacked",
 }: {
